Added a fallback in case of php out of support
So we were thinking that what if all of sudden we get married then she says she will only support drupal 11 and drupal team says we only want php8.4 and drupal 10 has a security cve and a upstream patch in a 8.13.x minor version ,in that case, we wont need to pay alimony anymore. this is the dockerfile to just build the php as apache and rest does not really matter as they built against base debian image and update against those like apache or mysql client isnt gonna be huge change. still opcache and extensions arent fully working. which we will fix later. at least a base working image template is here.

Signed-off-by: aless <38411333+alessblaze@users.noreply.github.com>