we build it all from source now.
what if there is a cve in apache or an rce in mysql client and the bitnami dude is sleeping or maybe theres a storm, or the dude forgot to change a small config change which failed the CI and job is stuck in pipeline or the guy is sad because of some reason and doesn't want to fix it immediately and they have a maintainance window to check it while CVE or RCE is present. maybe the dude it in toilet so it can be a bit late. As we know these sort of things can happen. so we build all needed things from source. modsecurity3 isnt included as the connector is from 2017 idk if they have a backport fixed personal branch. i havent checked even modsecurity 2 is enabled or not but its included. 

Signed-off-by: aless <38411333+alessblaze@users.noreply.github.com>