hehe update php 8.4
this is updated due to personal reasons as a company we are leveraging php 8.4 becuase of modern super dooper ultra new featues that i do not know as im mainly c/c++/python dev. but i hope most of the modules will work yet or not because as far of my knowledge drupal 10.4 supports php 8.4 so drupal does not inherently have an api specific for plugin developments and most of plugins and themes rely on underlying php interpreter for lots of functions. hereby we updated it into a more active NEW branch to get best of latest php 8.4 as we think it will be more actively maintained than old stable 8.3. 

Signed-off-by: aless <38411333+alessblaze@users.noreply.github.com>