fixed path permission issue
we have added supervisord as entrypoint and spawning mysql process as a service in it along with a watchdog which would send sigterm to supervisord to exit the container if mysql fails. Mysql service would need to fail three times in order to exit the container as similar behavior in official bitnami container. we are using supercronic to run cron jobs to backup periodically. the backups are xbsteam full backups everyday at 2 am. the user is automatically added with reused functions from libmysql to grant initial privileges to the backup user. we could have used only root but this makes it simpler if user wants to use another tailor made container in the future. services can be stopped simply and backups restored in container. the container is rootless so supervisor also runs as /tmp/ socket for permissions issues. we also wrapped some bash scripts which later we will add a process watchdogs to look for fatherless child processes.

Signed-off-by: aless <38411333+alessblaze@users.noreply.github.com>